final.md
In life, there are times when you need to exchange Kyrgyz soms for another currency: you and your family gather to go on vacation abroad, or make a purchase, my project is currency exchange. transfers an unknown amount of money from the monetary currency of one country to the monetary currency of another. thereby facilitating the life of a tourist, student, labor migrant. considers without the right to make mistakes, instantly and honestly. and you can easily find out how many soms we need to get in other currencies that we need